Moved NSFont extension to its own file
